Payment_Service/Payment/requirements.txt
asgiref==3.8.1
autopep8==2.3.1
confluent-kafka==2.8.0
Django==5.1.2
django-silk==5.3.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gprof2dot==2024.6.6
psycopg2-binary==2.9.10
pycodestyle==2.12.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.17.0
sqlparse==0.5.3
